requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16,<3.17
PyJWT>=2.10,<2.11
python-dotenv>=1.1,<1.2
